// packages/shared/src/types/database/schema.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
      api_logs: {
        Row: {
          body: string | null
          created_at: string
          duration_ms: number | null
          error_code: string | null
          expires_at: string
          headers: Json | null
          id: string
          method: string
          response_body: string | null
          response_headers: Json | null
          status_code: number | null
          url: string
        }
        Insert: {
          body?: string | null
          created_at?: string
          duration_ms?: number | null
          error_code?: string | null
          expires_at?: string
          headers?: Json | null
          id?: string
          method: string
          response_body?: string | null
          response_headers?: Json | null
          status_code?: number | null
          url: string
        }
        Update: {
          body?: string | null
          created_at?: string
          duration_ms?: number | null
          error_code?: string | null
          expires_at?: string
          headers?: Json | null
          id?: string
          method?: string
          response_body?: string | null
          response_headers?: Json | null
          status_code?: number | null
          url?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never